feat(test): refactor improve test quality and change productId from L…

+++ src/main/java/com/ecommerce/pricing/application/service/PriceQueryFacade.java
@@ -7,9 +7,9 @@ import reactor.core.publisher.Mono;
 
 public interface PriceQueryFacade {
 
-  Flux<PriceResponse> getAllPriceByProduct(Long productId);
+  Flux<PriceResponse> getAllPriceByProduct(Integer productId);
 
   Mono<PriceResponse> getPriceForIdentifier(Long id);
 
-  Mono<PriceResponse> getApplicablePrice(Long productId, Integer brandId, LocalDateTime date);
+  Mono<PriceResponse> getApplicablePrice(Integer productId, Integer brandId, LocalDateTime date);
 }
